Added logout to main and balance windows.

+++ QT_frontend/DLLRestAPI/engine_class.cpp
@@ -48,7 +48,7 @@ void engine_class::account_info(int account_id) {
 }
 
 // Request to get info if the card is locked and how many attempts before the card will lock.
-void engine_class::card_status(int card_num) {
+void engine_class::card_status(long long card_num) {
     QString site_url = site_base_url + "/card/" + QString().setNum(card_num);
 
     QNetworkRequest request(site_url);
